We’ve fixed several bugs and made performance improvements to enhance your experience. Thank you for using our app!

+++ lib/models/vendor_models/post_models/payment_post_data.dart
@@ -1,6 +1,6 @@
 class PaymentPostData {
   bool? cardForm;
-  int? cardAmount;
+  double? cardAmount;
   Map<String, String>? paymentMethod;
 
   // Convert object to List<MapEntry<String, String>>
